All the changes were not invoking a change event. THis means that the parry checkboxes can never be kep in sync with the native checkboxes.
This necessitated adding a check thet I am not changing the current element.

On reload checkbox callbacks were not initialised. Meaning the checkboxes could not do a second action after the first was invoked as it reloads the page and prevents checking.

All callbacks should only ever invoke the onChange on the native input as the happy inputs may not even exist. This prevents inifinite looping.

The counter should be initialised to 0/10 so the select count does not just appear out of nowhere.

The counter used to show the Total including the check all button if any button but the check all button was clicked.

Actually initialise the group action state at all times. Any javascript could run before this callback and this would ignore the checked boxes.

The multirow selection did ot work due to
1) no change events
2) the last checkbox being change at the start of the function rather than the end.
3) we were taking the lastCheckbox and lastCheckbox???
4) rowindex is global not only in this part of the table.